refactor: email 패키지 구조 변경 domain -> dto
