+++ backend/src/main/java/com/hjjang/backend/domain/email/dto/MailResponse.java
@@ -1,6 +1,5 @@
 package com.hjjang.backend.domain.email.dto;
 
-import com.hjjang.backend.domain.email.domain.Email;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
